Convert plugin to bash, remove dependency on ruby
I'd like to be able to use this context on agents that don't have ruby
installed, so I thought I'd experiment with changing it to pure bash.

It's likely the job that uses this plugin will still need to have ruby
and the `gem` command available, but if the OIDC to API Token work is
done in bash-only then we have the option to run the gem publishing
command using the docker plugin.